change DEVELOPMENT var to 'bool' and add PROFILER
Since the use of the DEVELOPMENT environment variable was not really
that user friendly where you had to choose between setting it or
explicitly using 'noprofile'. Now the DEVELOPMENT environment varialbe
only enables features that are usefull for development, and must be
explicitly set to '1'.

Introducing PROFILER, we don't want to lose the functionality we already
had so here comes PROFILER. Since we are actually providing xhprof
functionality the naming is made consistent that way. The prepend.php is
moved to its own xhprof folder, no longer tideways. This already gives
us an opening to support the Tideways.io profiler in a later stage.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>