allow rootless operation
Added permissions script that makes the correct files and folders owned
and usable by www-data.

Allows running a container with --user www-data

moved dirty tricks from prepare to install since everything must be
prepared for rootless operation. That can't be modified again later.

Updated the location of the .checkfpm flag for the healthcheck. Since
only parts of the containers filesystem are available for www-data the
flag is now placed in /run/php/.

Extensions helper now also aborts when not running as root. One should
see this helper as something to use in development or to use as helper
to build your own production image with additional php extensions.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>